Add static cast to size_t

+++ core-writer/src/BufferedWriter.cpp
@@ -25,7 +25,7 @@ BufferedWriter::BufferedWriter(const std::string& filename, size_t total_frames,
 
 void BufferedWriter::cache_metadata(string name, uint64_t frame_index, const char* data)
 {
-    auto relative_frame_index = get_relative_data_index(frame_index);
+    auto relative_frame_index = get_relative_data_index(static_cast<const size_t>(frame_index));
     metadata_buffer->add_metadata_to_buffer(name, relative_frame_index, data);
 }
